requirements.txt
Pillow
discord.py
asyncio
requests
brainfuck-interpreter
imageio
aiofiles